+++ apps/ceo-front/src/components/FoodingLogo.tsx
@@ -1,0 +1,54 @@
+export const FoodingLogo = (props: React.ComponentPropsWithRef<'svg'>) => {
+  return (
+    <svg
+      width='156'
+      height='24'
+      viewBox='0 0 156 24'
+      fill='none'
+      xmlns='http://www.w3.org/2000/svg'
+      {...props}
+    >
+      <path
+        d='M7.15003 7.59915H4.85298V17.4617H1.49144V7.59915H0V5.01483H1.49144V4.38616C1.49144 2.85939 1.94931 1.73585 2.86504 1.01554C3.78077 0.297063 5.16209 -0.0401818 7.009 0.0038067V2.65411C6.20339 2.63945 5.64314 2.76592 5.32823 3.03535C5.0114 3.30478 4.85491 3.79232 4.85491 4.49613V5.013H7.15197V7.59732L7.15003 7.59915Z'
+        fill='#FF2B3D'
+      />
+      <path
+        d='M11.4233 17.1823C10.4349 16.6407 9.65806 15.873 9.09484 14.8813C8.53161 13.8895 8.25 12.7446 8.25 11.4466C8.25 10.1485 8.53917 9.00551 9.11941 8.01189C9.69775 7.02014 10.4897 6.25252 11.4933 5.71089C12.4969 5.16925 13.6157 4.89844 14.8518 4.89844C16.0879 4.89844 17.2068 5.16925 18.2104 5.71089C19.214 6.25252 20.0059 7.02014 20.5842 8.01189C21.1626 9.00364 21.4536 10.1485 21.4536 11.4466C21.4536 12.7446 21.1569 13.8877 20.5615 14.8813C19.9662 15.873 19.1629 16.6407 18.1518 17.1823C17.1406 17.7239 16.0085 17.9947 14.7573 17.9947C13.5061 17.9947 12.4099 17.7239 11.4214 17.1823H11.4233ZM16.4149 14.7786C16.9327 14.4965 17.3447 14.0726 17.6547 13.5085C17.9628 12.9445 18.1178 12.2572 18.1178 11.4485C18.1178 10.2438 17.7965 9.31555 17.1557 8.66745C16.515 8.01936 15.7307 7.69438 14.8046 7.69438C13.8785 7.69438 13.1017 8.01936 12.4761 8.66745C11.8505 9.31555 11.5386 10.2438 11.5386 11.4485C11.5386 12.6531 11.8429 13.5814 12.4534 14.2295C13.0639 14.8776 13.8312 15.2025 14.7592 15.2025C15.3451 15.2025 15.8989 15.0625 16.4149 14.7786Z'
+        fill='#FF2B3D'
+      />
+      <path
+        d='M26.2827 17.1823C25.2942 16.6407 24.5174 15.873 23.9542 14.8813C23.391 13.8895 23.1094 12.7446 23.1094 11.4466C23.1094 10.1485 23.3985 9.00551 23.9788 8.01189C24.5571 7.02014 25.349 6.25252 26.3526 5.71089C27.3562 5.16925 28.4751 4.89844 29.7112 4.89844C30.9473 4.89844 32.0662 5.16925 33.0697 5.71089C34.0733 6.25252 34.8653 7.02014 35.4436 8.01189C36.0219 9.00364 36.313 10.1485 36.313 11.4466C36.313 12.7446 36.0163 13.8877 35.4209 14.8813C34.8256 15.873 34.0223 16.6407 33.0112 17.1823C32 17.7239 30.8679 17.9947 29.6167 17.9947C28.3655 17.9947 27.2693 17.7239 26.2808 17.1823H26.2827ZM31.2742 14.7786C31.7921 14.4965 32.2041 14.0726 32.5141 13.5085C32.8222 12.9445 32.9771 12.2572 32.9771 11.4485C32.9771 10.2438 32.6558 9.31555 32.0151 8.66745C31.3744 8.01936 30.59 7.69438 29.6639 7.69438C28.7378 7.69438 27.961 8.01936 27.3354 8.66745C26.7098 9.31555 26.398 10.2438 26.398 11.4485C26.398 12.6531 26.7023 13.5814 27.3128 14.2295C27.9232 14.8776 28.6906 15.2025 29.6186 15.2025C30.2045 15.2025 30.7583 15.0625 31.2742 14.7786Z'
+        fill='#FF2B3D'
+      />
+      <path
+        d='M38.1794 8.14027C38.6909 7.16179 39.3892 6.40812 40.2723 5.88111C41.1573 5.35409 42.1412 5.09058 43.2241 5.09058C44.0474 5.09058 44.8335 5.26748 45.5803 5.62128C46.3271 5.97508 46.9227 6.44497 47.3652 7.0328V1.09375H50.6138V17.8071H47.3652V15.9552C46.9675 16.5725 46.4111 17.07 45.6942 17.4459C44.9772 17.8219 44.1464 18.0098 43.1998 18.0098C42.1319 18.0098 41.1554 17.7389 40.2704 17.1972C39.3855 16.6554 38.6872 15.8907 38.1775 14.9048C37.6659 13.919 37.4102 12.7857 37.4102 11.5051C37.4102 10.2244 37.6659 9.11874 38.1775 8.14027H38.1794ZM46.932 9.57389C46.6258 9.02477 46.2151 8.60278 45.696 8.30979C45.177 8.0168 44.6206 7.86939 44.025 7.86939C43.4295 7.86939 42.8805 8.01312 42.3783 8.29874C41.8742 8.58436 41.4672 9.00265 41.1535 9.55178C40.8399 10.1009 40.6849 10.7532 40.6849 11.5051C40.6849 12.2569 40.8418 12.9166 41.1535 13.4823C41.4653 14.0461 41.878 14.4792 42.3895 14.7814C42.9011 15.0836 43.4463 15.2328 44.025 15.2328C44.6038 15.2328 45.177 15.0854 45.696 14.7924C46.2151 14.4995 46.6258 14.0775 46.932 13.5283C47.2363 12.9792 47.3894 12.3195 47.3894 11.5511C47.3894 10.7827 47.2363 10.1249 46.932 9.57574V9.57389Z'
+        fill='#FF2B3D'
+      />
+      <path d='M57.2111 5.45312V18.0037H53.9102V5.45312H57.2111Z' fill='#FF2B3D' />
+      <path
+        d='M71.2277 6.29255C72.1592 7.22378 72.6268 8.52312 72.6268 10.1906V17.4491H69.4172V10.6134C69.4172 9.63112 69.1647 8.87484 68.6615 8.34635C68.1565 7.81787 67.4701 7.55362 66.5985 7.55362C65.7269 7.55362 65.0124 7.81787 64.5018 8.34635C63.9893 8.87484 63.733 9.63112 63.733 10.6134V17.4491H60.5234V5.07521H63.733V6.61693C64.1614 6.08115 64.7075 5.66019 65.3715 5.35585C66.0355 5.05152 66.7668 4.89844 67.5617 4.89844C69.0749 4.89844 70.2962 5.36314 71.2296 6.29437L71.2277 6.29255Z'
+        fill='#FF2B3D'
+      />
+      <path
+        d='M83.1089 5.45397C83.8258 5.82681 84.3895 6.3115 84.8021 6.90432V5.09977H88.0279V17.8509C88.0279 19.0235 87.7908 20.0712 87.3185 20.9921C86.8462 21.913 86.1368 22.6457 85.1904 23.1863C84.2439 23.7269 83.0996 23.9972 81.7574 23.9972C79.9578 23.9972 78.4812 23.5778 77.3294 22.7407C76.1776 21.9018 75.5261 20.761 75.373 19.3125H78.554C78.722 19.8904 79.0842 20.3527 79.6405 20.6957C80.1968 21.0387 80.8725 21.2102 81.6659 21.2102C82.5956 21.2102 83.3516 20.9325 83.9303 20.3769C84.5109 19.8214 84.8002 18.9788 84.8002 17.8528V15.8879C84.3877 16.4826 83.8202 16.9766 83.0959 17.3737C82.3716 17.7689 81.5427 17.9684 80.6131 17.9684C79.5453 17.9684 78.569 17.6943 77.6841 17.1463C76.7993 16.5982 76.1011 15.8245 75.5915 14.8272C75.08 13.8298 74.8242 12.6834 74.8242 11.3877C74.8242 10.0921 75.08 8.97359 75.5915 7.9837C76.103 6.9938 76.7955 6.23134 77.6729 5.69818C78.5503 5.16502 79.5303 4.89844 80.6131 4.89844C81.5576 4.89844 82.3902 5.08486 83.1071 5.4577L83.1089 5.45397ZM84.3447 9.43032C84.0386 8.87479 83.6279 8.44788 83.1089 8.15147C82.59 7.85507 82.0337 7.70593 81.4382 7.70593C80.8427 7.70593 80.2938 7.85134 79.7898 8.14029C79.2858 8.42924 78.8788 8.85241 78.5652 9.40795C78.2516 9.96348 78.0967 10.6234 78.0967 11.384C78.0967 12.1446 78.2535 12.812 78.5652 13.3843C78.877 13.9547 79.2895 14.3928 79.801 14.6986C80.3125 15.0043 80.8576 15.1553 81.4363 15.1553C82.015 15.1553 82.5881 15.0062 83.1071 14.7097C83.626 14.4133 84.0367 13.9864 84.3429 13.4309C84.6471 12.8754 84.8002 12.208 84.8002 11.4306C84.8002 10.6532 84.6471 9.98772 84.3429 9.43218L84.3447 9.43032Z'
+        fill='#FF2B3D'
+      />
+      <path
+        d='M54.1362 3.2654C53.7746 2.90131 53.5938 2.4481 53.5938 1.90577C53.5938 1.36343 53.7746 0.910216 54.1362 0.54613C54.4978 0.182043 54.9517 0 55.4995 0C56.0474 0 56.4994 0.182043 56.8628 0.54613C57.2245 0.910216 57.4053 1.36343 57.4053 1.90577C57.4053 2.4481 57.2245 2.90131 56.8628 3.2654C56.5012 3.62949 56.0456 3.81153 55.4995 3.81153C54.9535 3.81153 54.4978 3.62949 54.1362 3.2654Z'
+        fill='#FF2B3D'
+      />
+      <path
+        d='M153.373 16.1706C153.373 14.062 152.963 13.3902 150.745 13.3902H140.344C138.088 13.3902 137.678 14.062 137.678 16.2078V17.9433L137.683 18.3285C137.735 20.109 138.197 20.7202 140.138 20.759L140.344 20.7608H153.373V16.1706ZM149.254 15.984C149.608 15.984 149.701 16.0586 149.701 16.4505V18.1484H141.779V15.984H149.254ZM141.35 17.6634V16.4505C141.35 16.1076 141.407 16.0076 141.659 15.988L141.779 15.984V18.1484L141.659 18.1452C141.443 18.131 141.371 18.063 141.354 17.8073L141.35 17.6634ZM153.355 5.08657C153.355 3.91157 152.532 3.27929 151.583 3.241L151.491 3.23918C150.503 3.23918 149.627 3.87373 149.627 5.08657V12.7371H153.355V5.08657ZM155.591 12.7371C155.591 13.131 155.489 13.5009 155.311 13.8224C155.594 14.666 155.61 15.5349 155.61 16.1706V20.7608C155.61 21.9975 154.609 23 153.373 23H140.344C139.184 23 137.579 22.8768 136.488 21.6384C135.476 20.4901 135.441 18.9367 135.441 17.9433V16.2078C135.441 15.4096 135.464 14.2503 135.997 13.2298C135.977 13.2096 135.956 13.1894 135.936 13.1686C135.009 12.2107 134.789 10.9868 134.789 9.99425V9.65823H134.826C135.945 9.65823 136.299 8.8 136.299 8.03497C136.299 7.28858 135.963 6.41135 134.826 6.41135H134.789V5.40401C134.789 4.21615 135.264 3.14114 136.153 2.40386C136.986 1.71275 137.998 1.4665 138.89 1.4665C139.804 1.46651 140.809 1.72471 141.626 2.4188C142.487 3.15056 142.935 4.19903 142.935 5.3293V7.04585H147.036V9.28503H141.108L141.005 9.28066C140.799 9.26029 140.711 9.16893 140.699 8.94682L140.698 8.89325V5.3293C140.698 4.31703 139.963 3.74244 138.985 3.7075L138.89 3.70567C137.883 3.70567 137.026 4.26576 137.026 5.40401V9.99425L137.028 10.1236C137.066 11.4419 137.71 12.1773 139.263 12.1773H147.036C148.098 12.1773 148.434 11.4497 148.434 10.7221C148.434 10.0351 148.136 9.3305 147.135 9.28722L147.036 9.28503V7.04585C147.154 7.04585 147.272 7.0516 147.39 7.06116V5.08657C147.39 3.9012 147.838 2.81726 148.687 2.04779C149.503 1.30893 150.532 1 151.491 1C152.449 1 153.478 1.30893 154.294 2.04779C155.143 2.81726 155.591 3.9012 155.591 5.08657V12.7371Z'
+        fill='#FF2B3D'
+      />
+      <path
+        d='M134.269 17.2116C134.269 15.1085 132.918 13.6564 130.197 13.6564H122.614C119.873 13.6564 118.541 15.1457 118.541 17.2116L118.545 17.4018C118.623 19.2886 119.874 20.6747 122.361 20.7622L122.614 20.7665H130.197L130.448 20.7622C132.919 20.6768 134.186 19.3226 134.265 17.405L134.269 17.2116ZM123.552 16.188H129.277V18.2353H123.552L123.403 18.2299C122.681 18.1767 122.351 17.7351 122.351 17.2116C122.351 16.6882 122.681 16.2462 123.403 16.193L123.552 16.188ZM129.277 16.188C130.103 16.188 130.46 16.6533 130.46 17.2116L130.455 17.3149C130.413 17.8261 130.051 18.2353 129.277 18.2353V16.188ZM127.625 1.59581V3.82929H119.029V1.59581H127.625ZM136.034 8.01707C136.034 7.29587 135.717 6.45221 134.656 6.40012L134.551 6.39758H133.969V5.07617C133.969 3.90416 133.141 3.27349 132.185 3.2353L132.092 3.23348C131.098 3.23348 130.215 3.86642 130.215 5.07617V13.0795H133.969V9.6362H134.551L134.654 9.63365C135.632 9.58601 135.988 8.85975 136.029 8.16102L136.034 8.01707ZM136.766 9.97136C136.766 10.3872 136.823 10.7483 136.945 11.0507C136.724 11.2328 136.482 11.3859 136.222 11.5087V13.0795C136.222 13.548 136.076 13.9825 135.827 14.3416C136.309 15.2253 136.522 16.2177 136.522 17.2116C136.522 18.7424 136.013 20.2708 134.791 21.4027C133.585 22.5202 131.955 23 130.197 23H122.614C120.834 23 119.204 22.5049 118.003 21.3805C116.792 20.2464 116.289 18.7237 116.289 17.2116C116.289 16.3001 116.469 15.385 116.878 14.5536C116.794 14.4969 116.711 14.4372 116.631 14.3729C116.171 14.0046 115.805 13.5326 115.57 12.9864H115.782C116.908 12.9864 117.264 12.1304 117.264 11.3673C117.264 10.6228 116.927 9.7478 115.782 9.7478H115.72C115.83 9.54804 115.962 9.35461 116.119 9.17162C116.388 8.85693 116.7 8.60837 117.034 8.4184C115.97 7.69747 115.455 6.44682 115.41 5.45933L115.406 5.31828L115.409 5.17723C115.435 4.4657 115.662 3.6567 116.197 2.96411C116.823 2.15384 117.812 1.59581 119.029 1.59581V3.82929L118.941 3.83147C118.046 3.88021 117.659 4.6692 117.659 5.31828C117.659 5.96733 118.117 6.72109 118.947 6.76764L119.029 6.76982H121.394V7.21659C121.394 8.74281 120.455 9.91545 118.691 10.176C117.959 10.2877 117.527 10.7904 117.527 11.5349C117.527 12.484 118.334 13.0238 119.254 13.0239C121.037 13.0239 122.802 12.1305 123.89 10.6043C125.035 10.5671 125.729 11.5906 125.617 12.8934H129.108C129.183 9.67347 127.7 8.12882 124.885 8.09159C124.922 7.79379 124.941 7.51439 124.941 7.21659V6.76982H127.625C128.507 6.76982 128.995 5.98831 128.995 5.31828C128.995 4.66919 128.608 3.88017 127.713 3.83147L127.625 3.82929V1.59581C128.257 1.59581 128.827 1.74665 129.317 2.00478C130.131 1.29752 131.145 1 132.092 1C133.058 1 134.094 1.30814 134.915 2.04513C135.631 2.68739 136.062 3.5495 136.184 4.50654C136.414 4.6115 136.63 4.73986 136.83 4.89041C136.788 5.04302 136.766 5.21054 136.766 5.3928V9.97136Z'
+        fill='#FF2B3D'
+      />
+      <path
+        d='M106.304 5.60919C106.304 4.47095 105.414 3.79897 104.449 3.79897C103.447 3.79903 102.558 4.41503 102.558 5.60919V7.06444L102.555 7.54552C102.507 12.4414 101.693 15.0761 98.9445 17.0103L98.8691 17.0675C98.115 17.6743 98.166 18.7076 98.6607 19.268C99.1715 19.9211 100.401 20.0706 101.385 19.3989C103.277 18.2046 104.658 16.3198 105.471 13.6701C106.833 14.8083 107.23 16.8797 106.89 19.6974H110.522C110.863 14.4166 109.425 11.5428 106.133 10.2366C106.232 9.37132 106.288 8.44891 106.301 7.46935L106.304 7.04585V5.60919ZM115.479 5.08657C115.478 3.91157 114.644 3.27929 113.68 3.241L113.587 3.23918C112.584 3.23918 111.695 3.87373 111.695 5.08657V20.7608H115.479V13.017H116.159L116.264 13.0144C117.249 12.9667 117.607 12.2386 117.65 11.5381L117.654 11.3937C117.654 10.6707 117.335 9.82488 116.266 9.77267L116.159 9.77011H115.479V5.08657ZM117.749 7.84874C118.317 8.09529 118.805 8.48074 119.179 8.9891C119.434 9.33706 119.606 9.70336 119.72 10.0595C119.524 10.1197 119.314 10.1667 119.092 10.1994C118.354 10.3114 117.919 10.8154 117.919 11.5617C117.919 12.4925 118.697 13.03 119.599 13.0531C119.489 13.3121 119.346 13.5716 119.16 13.8206C118.788 14.318 118.306 14.6943 117.749 14.9369V20.7608C117.749 21.9975 116.732 23 115.479 23H111.695C110.878 23 110.163 22.5747 109.763 21.9365H106.89C106.241 21.9365 105.623 21.6625 105.192 21.1839C104.811 20.7599 104.609 20.2112 104.621 19.6496C104.02 20.2698 103.35 20.8167 102.609 21.2849L102.608 21.2842C100.926 22.3876 98.3042 22.4641 96.8721 20.6464C95.6804 19.1953 95.5994 16.6412 97.6092 15.1993L97.6262 15.1873L97.8168 15.0488C98.7449 14.3511 99.3005 13.5772 99.6668 12.5181C100.092 11.2898 100.288 9.60242 100.288 7.06444V5.60919C100.288 4.42001 100.753 3.33877 101.624 2.5788C102.456 1.85342 103.494 1.55982 104.449 1.55979L104.64 1.56417C106.608 1.65437 108.574 3.13594 108.574 5.60919V7.04585C108.574 7.69022 108.556 8.31699 108.521 8.92569C108.838 9.12031 109.139 9.33251 109.425 9.56238V5.08657C109.425 3.90121 109.88 2.81726 110.742 2.04779C111.57 1.30893 112.614 1 113.587 1C114.56 1 115.604 1.30893 116.432 2.04779C117.294 2.81726 117.749 3.9012 117.749 5.08657V7.84874Z'
+        fill='#FF2B3D'
+      />
+    </svg>
+  );
+};
